more flattening
